Clean up the qsim examples
- Renamed those example files to be more specific.

- Added qcor-qsim linkage to the qcor driver script.

- Added unit tests to compile those two example with qcor.

- Remove debug logging in the impl.
