+++ lib/qsim/impls/CMakeLists.txt
@@ -39,6 +39,6 @@ endif()
 install(TARGETS ${LIBRARY_NAME} DESTINATION ${CMAKE_INSTALL_PREFIX}/plugins)
 install(FILES ${HEADERS} DESTINATION include/qcor)
 
-# if (QCOR_BUILD_TESTS)
-# 	add_subdirectory(tests)
-# endif()
+if (QCOR_BUILD_TESTS)
+	add_subdirectory(tests)
+endif()
